chat/pyproject.toml
[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
known_third_party = "kedro"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/llm_db_interface -ra"""
pythonpath = [
  "./src",
]

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.poetry]
name = "llm-db-interface"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">3.9.7,<3.11"
google-cloud-aiplatform = "1.31.0"
langchain = "0.0.201"
tensorflow_hub = "0.13.0"
tensorflow_text = "^2.12.1"
streamlit = "^1.27.1"
streamlit-chat = "^0.1.1"

[tool.poetry.dev-dependencies]
black = "^22.12"
flake8 = "^6.0"
ipython = "^7"
isort = "^5"
jupyter = "^1.0"
jupyter-client = "^7"
jupyterlab = "^3"
notebook = "^6.5.2"
nbstripout = "^0.6"
pytest = "^7.2"
pytest-cov = "^4.0"
pytest-mock = "^3.10"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
